Cyphers/opmove (#4560)
* Move non-experimantal ops out of experimental

* Redirects for compatibility
