+++ src/ngraph/op/experimental/layers/interpolate.hpp
@@ -14,52 +14,4 @@
 // limitations under the License.
 //*****************************************************************************
 
-#pragma once
-
-#include "ngraph/op/op.hpp"
-
-namespace ngraph
-{
-    namespace op
-    {
-        typedef struct
-        {
-            AxisSet axes;
-            std::string mode;
-            bool align_corners = true;
-            bool antialias = false;
-            std::vector<size_t> pads_begin;
-            std::vector<size_t> pads_end;
-        } InterpolateAttrs;
-
-        namespace v0
-        {
-            /// \brief Layer which performs bilinear interpolation
-            class NGRAPH_API Interpolate : public Op
-            {
-            public:
-                static constexpr NodeTypeInfo type_info{"Interpolate", 0};
-                const NodeTypeInfo& get_type_info() const override { return type_info; }
-                Interpolate() = default;
-                /// \brief Constructs a Interpolate operation
-                ///
-                /// \param image	    Input image
-                /// \param output_shape Output shape of spatial axes
-                /// \param attrs        Interpolation attributes
-                Interpolate(const Output<Node>& image,
-                            const Output<Node>& output_shape,
-                            const InterpolateAttrs& attrs);
-
-                void validate_and_infer_types() override;
-
-                virtual std::shared_ptr<Node>
-                    clone_with_new_inputs(const OutputVector& new_args) const override;
-
-                const InterpolateAttrs& get_attrs() const { return m_attrs; }
-            private:
-                InterpolateAttrs m_attrs;
-            };
-        }
-        using v0::Interpolate;
-    }
-}
+#include "ngraph/op/interpolate.hpp"
