+++ src/ngraph/op/experimental/layers/reorg_yolo.hpp
@@ -14,38 +14,4 @@
 // limitations under the License.
 //*****************************************************************************
 
-#pragma once
-
-#include "ngraph/op/op.hpp"
-
-namespace ngraph
-{
-    namespace op
-    {
-        namespace v0
-        {
-            class NGRAPH_API ReorgYolo : public Op
-            {
-            public:
-                static constexpr NodeTypeInfo type_info{"ReorgYolo", 0};
-                const NodeTypeInfo& get_type_info() const override { return type_info; }
-                ReorgYolo() = default;
-                /// \brief Constructs a ReorgYolo operation
-                ///
-                /// \param input          Input
-                /// \param strides        Stride to reorganize input by
-                ReorgYolo(const Output<Node>& input, const Strides& strides);
-
-                void validate_and_infer_types() override;
-
-                virtual std::shared_ptr<Node>
-                    clone_with_new_inputs(const OutputVector& new_args) const override;
-
-                const Strides get_strides() const { return m_strides; }
-            private:
-                Strides m_strides;
-            };
-        }
-        using v0::ReorgYolo;
-    }
-}
+#include "ngraph/op/reorg_yolo.hpp"
