+++ src/ngraph/op/experimental/range.hpp
@@ -14,45 +14,4 @@
 // limitations under the License.
 //*****************************************************************************
 
-#pragma once
-
-#include "ngraph/node.hpp"
-#include "ngraph/op/op.hpp"
-
-namespace ngraph
-{
-    namespace op
-    {
-        namespace v0
-        {
-            /// \brief Range operation, analogous to `range()` in Python.
-            class NGRAPH_API Range : public Op
-            {
-            public:
-                static constexpr NodeTypeInfo type_info{"Range", 0};
-                const NodeTypeInfo& get_type_info() const override { return type_info; }
-                /// \brief Constructs an unitialized range operation.
-                Range() = default;
-
-                /// \brief Constructs a range operation.
-                ///
-                /// \param start The tensor producing the start value. Must be a scalar of integer
-                ///              element type, and same element type as `stop` and `step`.
-                /// \param stop The tensor producing the stop value. Must be a scalar of integer
-                ///             element type, and same element type as `start` and `step`.
-                /// \param step The tensor producing the step value. Must be a scalar of integer
-                ///             element type, and same element type as `start` and `stop`.
-                Range(const Output<Node>& start,
-                      const Output<Node>& stop,
-                      const Output<Node>& step);
-
-                bool visit_attributes(AttributeVisitor& visitor) override;
-                void validate_and_infer_types() override;
-
-                virtual std::shared_ptr<Node>
-                    clone_with_new_inputs(const OutputVector& new_args) const override;
-            };
-        }
-        using v0::Range;
-    }
-}
+#include "ngraph/op/range.hpp"
