+++ src/ngraph/op/interpolate.hpp
@@ -1,0 +1,65 @@
+//*****************************************************************************
+// Copyright 2017-2020 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//*****************************************************************************
+
+#pragma once
+
+#include "ngraph/op/op.hpp"
+
+namespace ngraph
+{
+    namespace op
+    {
+        typedef struct
+        {
+            AxisSet axes;
+            std::string mode;
+            bool align_corners = true;
+            bool antialias = false;
+            std::vector<size_t> pads_begin;
+            std::vector<size_t> pads_end;
+        } InterpolateAttrs;
+
+        namespace v0
+        {
+            /// \brief Layer which performs bilinear interpolation
+            class NGRAPH_API Interpolate : public Op
+            {
+            public:
+                static constexpr NodeTypeInfo type_info{"Interpolate", 0};
+                const NodeTypeInfo& get_type_info() const override { return type_info; }
+                Interpolate() = default;
+                /// \brief Constructs a Interpolate operation
+                ///
+                /// \param image	    Input image
+                /// \param output_shape Output shape of spatial axes
+                /// \param attrs        Interpolation attributes
+                Interpolate(const Output<Node>& image,
+                            const Output<Node>& output_shape,
+                            const InterpolateAttrs& attrs);
+
+                void validate_and_infer_types() override;
+
+                virtual std::shared_ptr<Node>
+                    clone_with_new_inputs(const OutputVector& new_args) const override;
+
+                const InterpolateAttrs& get_attrs() const { return m_attrs; }
+            private:
+                InterpolateAttrs m_attrs;
+            };
+        }
+        using v0::Interpolate;
+    }
+}
