+++ src/ngraph/ops.hpp
@@ -44,10 +44,12 @@
 #include "ngraph/op/cos.hpp"
 #include "ngraph/op/cosh.hpp"
 #include "ngraph/op/crop_and_resize.hpp"
+#include "ngraph/op/ctc_greedy_decoder.hpp"
 #include "ngraph/op/cum_sum.hpp"
 #include "ngraph/op/deformable_convolution.hpp"
 #include "ngraph/op/deformable_psroi_pooling.hpp"
 #include "ngraph/op/dequantize.hpp"
+#include "ngraph/op/detection_output.hpp"
 #include "ngraph/op/divide.hpp"
 #include "ngraph/op/dot.hpp"
 #include "ngraph/op/embedding_lookup.hpp"
@@ -61,24 +63,10 @@
 #include "ngraph/op/experimental/dyn_replace_slice.hpp"
 #include "ngraph/op/experimental/dyn_slice.hpp"
 #include "ngraph/op/experimental/generate_mask.hpp"
-#include "ngraph/op/experimental/layers/ctc_greedy_decoder.hpp"
-#include "ngraph/op/experimental/layers/detection_output.hpp"
-#include "ngraph/op/experimental/layers/interpolate.hpp"
-#include "ngraph/op/experimental/layers/prior_box.hpp"
-#include "ngraph/op/experimental/layers/prior_box_clustered.hpp"
-#include "ngraph/op/experimental/layers/proposal.hpp"
-#include "ngraph/op/experimental/layers/psroi_pooling.hpp"
-#include "ngraph/op/experimental/layers/region_yolo.hpp"
-#include "ngraph/op/experimental/layers/reorg_yolo.hpp"
-#include "ngraph/op/experimental/layers/roi_pooling.hpp"
 #include "ngraph/op/experimental/quantized_conv_bias.hpp"
 #include "ngraph/op/experimental/quantized_conv_relu.hpp"
 #include "ngraph/op/experimental/quantized_dot_bias.hpp"
 #include "ngraph/op/experimental/random_uniform.hpp"
-#include "ngraph/op/experimental/range.hpp"
-#include "ngraph/op/experimental/shape_of.hpp"
-#include "ngraph/op/experimental/tile.hpp"
-#include "ngraph/op/experimental/transpose.hpp"
 #include "ngraph/op/floor.hpp"
 #include "ngraph/op/floor_mod.hpp"
 #include "ngraph/op/fused/batch_mat_mul_transpose.hpp"
@@ -122,6 +110,7 @@
 #include "ngraph/op/greater.hpp"
 #include "ngraph/op/greater_eq.hpp"
 #include "ngraph/op/group_conv.hpp"
+#include "ngraph/op/interpolate.hpp"
 #include "ngraph/op/less.hpp"
 #include "ngraph/op/less_eq.hpp"
 #include "ngraph/op/log.hpp"
@@ -143,29 +132,38 @@
 #include "ngraph/op/parameter.hpp"
 #include "ngraph/op/passthrough.hpp"
 #include "ngraph/op/power.hpp"
+#include "ngraph/op/prior_box.hpp"
+#include "ngraph/op/prior_box_clustered.hpp"
 #include "ngraph/op/product.hpp"
+#include "ngraph/op/proposal.hpp"
+#include "ngraph/op/psroi_pooling.hpp"
 #include "ngraph/op/quantize.hpp"
 #include "ngraph/op/quantized_convolution.hpp"
 #include "ngraph/op/quantized_dot.hpp"
+#include "ngraph/op/range.hpp"
 #include "ngraph/op/recv.hpp"
 #include "ngraph/op/reduce_logical_and.hpp"
 #include "ngraph/op/reduce_logical_or.hpp"
 #include "ngraph/op/reduce_mean.hpp"
 #include "ngraph/op/reduce_prod.hpp"
 #include "ngraph/op/reduce_sum.hpp"
+#include "ngraph/op/region_yolo.hpp"
 #include "ngraph/op/relu.hpp"
+#include "ngraph/op/reorg_yolo.hpp"
 #include "ngraph/op/replace_slice.hpp"
 #include "ngraph/op/reshape.hpp"
 #include "ngraph/op/result.hpp"
 #include "ngraph/op/reverse.hpp"
 #include "ngraph/op/reverse_sequence.hpp"
 #include "ngraph/op/roi_align.hpp"
+#include "ngraph/op/roi_pooling.hpp"
 #include "ngraph/op/round.hpp"
 #include "ngraph/op/scatter_add.hpp"
 #include "ngraph/op/scatter_elements_update.hpp"
 #include "ngraph/op/scatter_nd_add.hpp"
 #include "ngraph/op/select.hpp"
 #include "ngraph/op/send.hpp"
+#include "ngraph/op/shape_of.hpp"
 #include "ngraph/op/sigmoid.hpp"
 #include "ngraph/op/sign.hpp"
 #include "ngraph/op/sin.hpp"
@@ -181,7 +179,9 @@
 #include "ngraph/op/tan.hpp"
 #include "ngraph/op/tanh.hpp"
 #include "ngraph/op/tensor_iterator.hpp"
+#include "ngraph/op/tile.hpp"
 #include "ngraph/op/topk.hpp"
+#include "ngraph/op/transpose.hpp"
 #include "ngraph/op/util/attr_types.hpp"
 #include "ngraph/op/variadic_split.hpp"
 #include "ngraph/op/xor.hpp"
